% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needmining.R
\docType{package}
\name{needmining-package}
\alias{needmining-package}
\alias{needmining}
\title{Functions for a simple needmining implementation}
\description{
needmining provides the basic functionality to
download social media data from Twitter and 
semi automatically classify the data regarding
user needs
}
