% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTweetsStopwords.R
\name{removeTweetsStopwords}
\alias{removeTweetsStopwords}
\title{Remove Tweets containing stopwords}
\usage{
removeTweetsStopwords(tweetMessages, stopWords)
}
\arguments{
\item{tweetMessages}{a dataframe containing the Tweet messages}

\item{stopWords}{a string containing stopwords separated by ';'}
}
\value{
a filtered data frame
}
\description{
\code{removeTweetsStopwords} removes Tweets containing stopwords
}
\details{
This function removes Tweets containing stopwords from a list of Twitter messages.
}
\examples{
stopWords <- "review;giveaway;save;deal;win;won;price;launch;news;gift;announce;
 			 reveal;sale;http;buy;bought;purchase;sell;sold;invest;discount;
			coupon;ship;giving away"
data(NMTrainingData)
filteredTweets <- removeTweetsStopwords(NMTrainingData, stopWords)
}
\author{
Dorian Proksch <dorian.proksch@hhl.de>
}
