% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveNeedminingFile.R
\name{saveNeedminingFile}
\alias{saveNeedminingFile}
\title{Save Tweet file}
\usage{
saveNeedminingFile(filename, tweetMessages)
}
\arguments{
\item{filename}{The filename to save to}

\item{tweetMessages}{An object containing the Twitter messages}
}
\description{
\code{saveNeedminingFile} saves a dataframe
created by the needmining package
to a file
}
\details{
This function saves a dataframe created by the needmining package to a file
}
\examples{
data(NMTrainingData)
saveNeedminingFile(filename=file.path(tempdir(), "NMTrainingData.csv"),
NMTrainingData)
}
\author{
Dorian Proksch <dorian.proksch@hhl.de>
}
